+++ packages/triple-web/src/contexts/session.tsx
@@ -3,7 +3,6 @@ import {
   PropsWithChildren,
   SetStateAction,
   createContext,
-  useContext,
   useState,
 } from 'react'
 
@@ -15,16 +14,6 @@ export const SessionStateContext = createContext<SessionValue | undefined>(
 export const SessionUpdaterContext = createContext<
   Dispatch<SetStateAction<SessionValue>> | undefined
 >(undefined)
-
-export function useSession() {
-  const context = useContext(SessionStateContext)
-
-  if (context === undefined) {
-    throw new Error('SessionContext가 없습니다.')
-  }
-
-  return context
-}
 
 export interface SessionProviderProps extends PropsWithChildren {
   initialSession: SessionValue
