+++ packages/triple-web/src/contexts/user-agent.tsx
@@ -1,17 +1,7 @@
-import { createContext, useContext } from 'react'
+import { createContext } from 'react'
 
 import { type UserAgentValue } from '../types'
 
 export const UserAgentContext = createContext<UserAgentValue | undefined>(
   undefined,
 )
-
-export function useUserAgent() {
-  const context = useContext(UserAgentContext)
-
-  if (context === undefined) {
-    throw new Error('UserAgentContext가 없습니다.')
-  }
-
-  return context
-}
