+++ packages/triple-web/src/hooks/client-app/use-client-app.ts
@@ -1,1 +1,13 @@
-export { useClientApp } from '../../contexts'
+import { useContext } from 'react'
+
+import { ClientAppContext } from '../../contexts'
+
+export function useClientApp() {
+  const context = useContext(ClientAppContext)
+
+  if (context === undefined) {
+    throw new Error('ClientAppContext가 없습니다.')
+  }
+
+  return context
+}
